"Removed Duplicates"

+++ src/main/java/com/backend/Retrospect/room/repository/IRoomRepository.java
@@ -1,9 +1,10 @@
 package com.backend.Retrospect.room.repository;
 
 
-import com.backend.Retrospect.room.entiry.RoomEntity;
+import com.backend.Retrospect.room.entity.RoomEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IRoomRepository extends JpaRepository<RoomEntity , Long> {
+public interface IRoomRepository extends JpaRepository<RoomEntity,Integer>
+{
 
 }
